API/ENH: Add image database step, postproc module
This is a big one.

We add a processing step: load all images in parallel (we use
multiprocessing) into a in-memory db (a dict) and resize to the
dimensions used for the NN model. It turns out that loading many images
from disk (even < 10 MB/image, even with an SSD) is slow. With the
images in memory, the fingerprints loop actually loads all CPU cores
~100% with TensorFlow w/o waiting for IO. Functions such as
fingerprint(s) which used to get a file name or a list of those now work
with (the dict of) image arrays.

Rename imagecluster.py -> calc.py

Add postproc.py and move make_links() there. Add function to plot a grid
of images arranged into clusters. We build a numpy array out of images
and use only one imshow() call .. cool eh? We use the in-memory images
here as well. Else, plotting would be painfully slow.