BUG/API: rename *elem* -> *csize*, use >= when checking min_csize
Change all variable names related to "the size of clusters" from *elem*
(meaning number of elements) to "csize" for "cluster size".

Set the default min_csize to 1 instead of 0 and use >= to compare, which
is what the docs already say. Default behavior is, however unchanged:
Clusters of csize at least 2 are reported be default.